+++ DcsBiosCommunicator/Configuration/OutputInteger.cs
@@ -4,7 +4,7 @@ public record OutputInteger : BiosOutput
 {
     public static string OutputType => "integer";
 
-    public int Mask { get; set; }
-    public int MaxValue { get; set; }
-    public int ShiftBy { get; set; }
+    public ushort Mask { get; set; }
+    public ushort MaxValue { get; set; }
+    public byte ShiftBy { get; set; }
 }
