+++ DcsBiosCommunicator/DataParsers/DataParser.cs
@@ -2,17 +2,17 @@
 
 public abstract class DataParser<T>
 {
-    public int Address { get; }
+    public ushort Address { get; }
 
     public string BiosCode { get; }
 
     public T CurrentValue { get; protected set; } = default!;
 
-    protected DataParser(in int address, in string biosCode)
+    protected DataParser(in ushort address, in string biosCode)
     {
         Address = address;
         BiosCode = biosCode;
     }
 
-    public abstract void AddData(in int address, in int data);
+    public abstract void AddData(in ushort address, in ushort data);
 }
